+++ dbo/Stored Procedures/DsmrApi_ComputeDeviceCapabilities.sql
@@ -4,13 +4,13 @@ AS
 BEGIN
 	SELECT CONVERT(BIT,
 		CASE
-			WHEN MIN([EnergyProduction]) - MAX([EnergyProduction]) != 0 THEN 1
+			WHEN MIN([EnergyProduction]) - MAX([EnergyProduction]) != 0.0 THEN 1
 			ELSE 0
 		END) AS [HasSolarCells],
 
 	   CONVERT(BIT,
 		CASE
-			WHEN MIN([GasUsage]) - MAX([GasUsage]) != 0 THEN 1
+			WHEN MIN([GasUsage]) - MAX([GasUsage]) != 0.0 THEN 1
 			ELSE 0
 		END) AS [HasGasMeter]
 	FROM [dbo].[DataPoints]